+++ examples/minimal.rs
@@ -13,7 +13,8 @@ fn main() {
     opt.fontdb.load_system_fonts();
     opt.fontdb.set_generic_families();
 
-    let rtree = usvg::Tree::from_file(&args[1], &opt).unwrap();
+    let svg_data = std::fs::read(&args[1]).unwrap();
+    let rtree = usvg::Tree::from_data(&svg_data, &opt).unwrap();
 
     let pixmap_size = rtree.svg_node().size.to_screen_size();
     let mut pixmap = tiny_skia::Pixmap::new(pixmap_size.width(), pixmap_size.height()).unwrap();